+++ kannich-tools/CHANGES.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2026-02-20
+## [0.6.0] - 2026-02-20
 ### Added
-- Supports the new `Tool` interface of Kannich.
+- `Fs` tool now has support for `chmod`.
+- `Fs` tool now has a new method `getParent` which returns the parent directory of a given path.
+- `Docker` tool now has a new method `tagAndPush` which tags and pushes a given image.
 
-## [0.1.0] - 2026-02-19
+## [0.5.0] - 2026-02-05
 - Initial release